% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cellTypeIdentification.R
\name{classifyTarget}
\alias{classifyTarget}
\alias{classifyTarget,matrix-method}
\alias{classifyTarget,Matrix-method}
\alias{classifyTarget,SummarizedExperiment-method}
\title{Classify target cells}
\usage{
classifyTarget(target, ...)

\S4method{classifyTarget}{matrix}(target, d.list, return.probs = F)

\S4method{classifyTarget}{Matrix}(target, d.list, return.probs = F)

\S4method{classifyTarget}{SummarizedExperiment}(target, assay = "counts", d.list, return.probs = F)
}
\arguments{
\item{target}{an object inheriting from \code{\link{SummarizedExperiment}}
(such as
\code{\link{SingleCellExperiment}}). Alternatively, a matrix or matrix-like
object (such as a sparse \code{\link{Matrix}}) of non-negative integer 
counts, with rownames indicating genes. Genes must be in the format of Ensembl IDs.}

\item{...}{for the generic, additional arguments pass to object-specific methods.}

\item{d.list}{an object returned from \code{\link{trainAllReference}}}

\item{return.probs}{if TRUE, the output will be the probabilities that each cell
belongs to each cell-type label under consideration. If FALSE (default), the
output will be the highest probability cell-type label for each cell.}

\item{assay}{a string or integer specifying which assay contains the count
data (default = 'counts'). Ignored if \code{target} is a matrix-like 
object.}
}
\value{
The inputted \code{\link{SummarizedExperiment}} object with the classification
  results appended to the colData. If return.probs = TRUE, the probabilities of each
  cell's label will be found under columns named after each cell-type. If 
  return.probs = FALSE, the highest probability cell-type label will be found under
  a column named predicted_labels.
}
\description{
Probabilistically predicts cell-type labels for unknown target cells, 
  based on reference data fitted to a latent states model.
}
\details{
We compute the probability of each unknown target cell belonging to each
  cell-type label based on the fits of the reference data to a latent states model.
  We assume independence of genes, and multiply together the probabilities of each
  observed count for each gene in an unknown cell arising under each cell-type label. 
  Bayes rule then allows us to compute the probability of each cell-type label.
}
\examples{

}
\references{
Grabski IN and Irizarry RA (2020). A probabilistic gene barcode for annotation
of cell-types from single cell RNA-seq data. \emph{bioRxiv}.
}
