% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.R
\name{compute_size_factors}
\alias{compute_size_factors}
\title{compute_size_factors}
\usage{
compute_size_factors(m, mod = c("binomial", "poisson", "geometric"))
}
\arguments{
\item{m}{a matrix of integer count values.}

\item{mod}{a character specifying the model type to be used for calculatind
size factors.}
}
\value{
A vector of size factors with length equal to the number of columns
  of \code{m}.
}
\description{
This function computes column-wise size factors as appropriate
  to a particular, discrete model.
}
