% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/devianceResiduals.R
\name{devianceResiduals}
\alias{devianceResiduals}
\alias{devianceResiduals,SingleCellExperiment-method}
\alias{devianceResiduals,SummarizedExperiment-method}
\alias{devianceResiduals,matrix-method}
\title{Deviance Residuals}
\usage{
devianceResiduals(object, ...)

\S4method{devianceResiduals}{SingleCellExperiment}(
  object,
  assay = 1,
  fam = c("binomial", "poisson", "geometric"),
  recalcSizeFactors = TRUE
)

\S4method{devianceResiduals}{SummarizedExperiment}(
  object,
  assay = 1,
  fam = c("binomial", "poisson", "geometric"),
  recalcSizeFactors = TRUE
)

\S4method{devianceResiduals}{matrix}(
  object,
  assay = 1,
  fam = c("binomial", "poisson", "geometric"),
  recalcSizeFactors = TRUE
)
}
\arguments{
\item{object}{A \code{\link{SingleCellExperiment}} or
\link{SummarizedExperiment} object. Alternatively,  a matrix of integer
counts.}

\item{...}{for the generic, additional arguments to pass to object-specific
methods.}

\item{assay}{a character or integer specifying which assay to use for GLM-PCA
(default = 1). Ignored if \code{object} is a matrix.}

\item{fam}{a character specifying the model type to be used for calculatind
deviance residuals.}

\item{recalcSizeFactors}{logical, whether column (cell) size factors should
be calculated according to the specified \code{fam} value. If \code{FALSE},
existing size factors are used.}
}
\value{
The original \code{SingleCellExperiment} or
  \code{SummarizedExperiment} object with the deviance residuals, p-values,
  and q-values added to the \code{colData}. If the input was a matrix, output
  is a \code{data.frame} with these columns.
}
\description{
This function computes deviance residuals for count-based
  features. Deviance residuals may be used in place of variance for feature
  selection.
}
