% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cellType.R
\name{fitType}
\alias{fitType}
\alias{fitType,SummarizedExperiment-method}
\title{Fit reference cell type model}
\usage{
fitType(reference, ...)

\S4method{fitType}{SummarizedExperiment}(reference, iters = 5000)
}
\arguments{
\item{reference}{A \code{\link{SummarizedExperiment}} object containing the
reference single-cell data.}

\item{...}{for the generic, additional arguments to pass to object-specific
methods.}

\item{iters}{The number of iterations used in MCMC. Defaults to 5000.}
}
\value{
A \code{data.frame}, where the first column contains the total number
  of counts of each gene observed across the entire dataset and the second
  column contains the mean parameter for each gene based on the model fit.
}
\description{
This function fits a zero-inflated mixture model to reference
  single-cell RNA-seq data consisting of entirely one cell type.
}
