% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cellTypeIdentification.R
\name{getBarcode}
\alias{getBarcode}
\alias{getBarcode,ANY-method}
\title{Get probabilistic barcode for each cell-type}
\usage{
getBarcode(d.list)

\S4method{getBarcode}{ANY}(d.list)
}
\arguments{
\item{d.list}{an object returned from \code{\link{trainAllReference}}}
}
\value{
A dataframe indicating the probability that each gene is on in each cell-type.
}
\description{
Defines probabilistic barcode for each cell-type in reference data
  by converting each cell-type's fit to the latent states model into a more
  interpretable dataframe of probabilities that each gene is on.
}
\details{
Fitting each cell-type to the latent states model results in probabilities
  of each gene in each cell-type belonging to one of three latent states. This method
  takes these fits and computes just the probability that each gene is on (expressed)
  in each cell-type, which is a more easily interpretable summary of the model fit.
}
\examples{

}
\references{
Grabski IN and Irizarry RA (2020). A probabilistic gene barcode for annotation
of cell-types from single cell RNA-seq data. \emph{bioRxiv}.
}
