% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/glmpca.R
\name{GLMPCA}
\alias{GLMPCA}
\alias{GLMPCA,SummarizedExperiment-method}
\alias{GLMPCA,matrix-method}
\alias{GLMPCA,Matrix-method}
\title{Generalized principal components analysis for non-normally distributed
  data}
\usage{
GLMPCA(object, ...)

\S4method{GLMPCA}{SummarizedExperiment}(object, assay = 1, L, ...)

\S4method{GLMPCA}{matrix}(object, assay = 1, L, ...)

\S4method{GLMPCA}{Matrix}(object, assay = 1, L, ...)
}
\arguments{
\item{object}{A \code{\link{SingleCellExperiment}} or
\link{SummarizedExperiment} object. Alternatively, a matrix of integer
counts.}

\item{...}{further arguments passed to \code{\link[glmpca]{glmpca}}}

\item{assay}{a character or integer specifying which assay to use for GLM-PCA
(default = 1). Ignored if \code{object} is a matrix.}

\item{L}{the desired number of latent dimensions (integer).}
}
\value{
The original \code{SingleCellExperiment} or
  \code{SummarizedExperiment} object with the GLM-PCA results added to the
  \code{metadata} slot. If the original input was a
  \code{SingleCellExperiment}, then a new \code{reducedDim} element called
  \code{"GLMPCA"} will be added, representing the GLM-PCA \code{factors}. If
  the input was a matrix, output matches that of
  \code{\link[glmpca]{glmpca}}.
}
\description{
This function implements the GLM-PCA dimensionality reduction
  method for high-dimensional count data. This is a wrapper for
  \code{\link[glmpca]{glmpca}}.
}
\examples{
ncells <- 100
u <- matrix(rpois(20000, 5), ncol=ncells)
sce <- SingleCellExperiment::SingleCellExperiment(assays=list(counts=u))
GLMPCA(sce, L = 2)

}
