% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTypeIdentification.R
\docType{data}
\name{params_EM_81020}
\alias{params_EM_81020}
\alias{params}
\title{Cell type model parameters}
\description{
Gene-specific parameters used for cell type prediction.
}
\examples{
data(params_EM_81020)

}
\references{
Grabski IN and Irizarry RA (2020). A probabilistic gene barcode for annotation
of cell-types from single cell RNA-seq data. \emph{bioRxiv}.
}
