% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cellTypeIdentification.R
\name{trainAllReference}
\alias{trainAllReference}
\alias{trainAllReference,matrix-method}
\alias{trainAllReference,Matrix-method}
\alias{trainAllReference,SummarizedExperiment-method}
\title{Train reference data to latent states model}
\usage{
trainAllReference(data, ...)

\S4method{trainAllReference}{matrix}(data, labels)

\S4method{trainAllReference}{Matrix}(data, labels)

\S4method{trainAllReference}{SummarizedExperiment}(data, assay = "counts", slot = "labels")
}
\arguments{
\item{data}{an object inheriting from \code{\link{SummarizedExperiment}}
(such as
\code{\link{SingleCellExperiment}}). Alternatively, a matrix or matrix-like
object (such as a sparse \code{\link{Matrix}}) of non-negative integer 
counts, with rownames indicating genes. Genes must be in the format of Ensembl IDs.}

\item{...}{for the generic, additional arguments pass to object-specific methods.}

\item{labels}{a vector of characters specifying the cell-type labels for
each cell. Ignored if \code{data} inherits from \code{\link{SummarizedExperiment}}.}

\item{assay}{a string or integer specifying which assay contains the count
data (default = 'counts'). Ignored if \code{data} is a matrix-like 
object.}

\item{slot}{a string or integer specifying which column of \code{colData} 
contains the character cell-type labels for each cell (default = 'labels'). 
Ignored if \code{data} is a matrix-like object.}
}
\value{
A list of dataframes indicating the cell-type-specific mixing proportions
  for the first two latent states (off-low and off-high) for each gene in
  each cell-type.
}
\description{
Fits latent states model to each cell-type present in the provided
  reference data by computing cell-type-specific mixing proportions for each
  of the three latent states (off-low, off-high, and on).
}
\details{
We define three latent states (off-low, indicating unexpressed genes
  with very low counts; off-high, indicating unexpressed genes with higher counts;
  and on, indicating expression) for each gene in human single-cell data. For
  each cell-type in the provided reference data, we learn the cell-type-specific
  mixing proportions among these three states for each gene. If there are genes
  present in the reference data that are not among the genes in our database,
  they are discarded. The object returned here can be passed to 
  \code{\link{getBarcode}} to obtain a more interpretable summary of the 
  probability that each gene is on in each cell-type, or to \code{\link{classifyTarget}}
  to label unknown cells.
}
\examples{

}
\references{
Grabski IN and Irizarry RA (2020). A probabilistic gene barcode for annotation
of cell-types from single cell RNA-seq data. \emph{bioRxiv}.
}
