% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.R
\name{compute_size_factors}
\alias{compute_size_factors}
\title{Compute size factors}
\usage{
compute_size_factors(m, fam = c("binomial", "poisson"))
}
\arguments{
\item{m}{a matrix or sparse \code{\link{Matrix}} of integer count values.}

\item{fam}{a string specifying the model type to be used for calculating
size factors.}
}
\value{
A vector of size factors with length equal to the number of columns
  of \code{m}.
}
\description{
Computes a size factor for each observation (column) of a 
  count data matrix based on an approximate multinomial model.
}
\details{
Both fam options are approximations to a multinomial model. Size factors
  for binomial are simply the column sums (total counts for each sample). For
  Poisson, the size factors are given by the column sums after rescaling to
  have geometric mean of one. This improves numerical stability.
}
