% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_size_factors}
\alias{compute_size_factors}
\title{Compute size factors}
\usage{
compute_size_factors(m)
}
\arguments{
\item{m}{a matrix-like object (such as sparse \code{\link{Matrix}} or
\code{\link{DelayedArray}}) of integer count values.}

\item{fam}{a string specifying the model type to be used for calculating
size factors. Must be either 'binomial' or 'poisson'.}
}
\value{
A vector of size factors with length equal to the number of columns
  of \code{m}.
}
\description{
Computes a size factor for each observation (column) of a
  count data matrix based on an approximate multinomial model.
}
\details{
Both fam options are approximations to a multinomial model. Size
  factors for binomial are simply the column sums (total counts for each
  sample). For Poisson, the size factors are given by the column sums after
  rescaling to have geometric mean of one. This improves numerical stability.
}
\examples{
ncells <- 100
u <- matrix(rpois(20000, 5), ncol=ncells)
compute_size_factors(u)

}
