% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cellType.R
\name{predictType}
\alias{predictType}
\alias{predictType,SummarizedExperiment-method}
\title{Predict cell type}
\usage{
predictType(target, ...)

\S4method{predictType}{SummarizedExperiment}(target, reference_dfs, types)
}
\arguments{
\item{target}{blah}

\item{...}{for the generic, additional arguments to pass to object-specific
methods.}

\item{reference_dfs}{}

\item{types}{}
}
\value{

}
\description{

}
